Created an emulate method ready for the future + Removed bad operation I forgot about
An emulate method was created as it is possible for the hardware to run
instructions on the Z80 during an interrupt. Majority of the code from
the "processOpcode" method has been moved to the "emulate" method. Also
I forgot to remove the duplicate operation in the last commit. Although
its inner contents was removed. So that has been removed on this commit
also.